[refactor] fix warning "Implicit cast from 'long' to 'int' in compound assignment can be lossy"

+++ flying-saucer-core/src/main/java/org/xhtmlrenderer/render/InlineLayoutBox.java
@@ -329,11 +329,11 @@ public final class InlineLayoutBox extends Box implements InlinePaintable {
             }
         }
         if (marginRight > 0) {
-            result.width += marginRight;
+            result.width += (int) marginRight;
         }
         if (marginLeft > 0) {
-            result.x -= marginLeft;
-            result.width += marginLeft;
+            result.x -= (int) marginLeft;
+            result.width += (int) marginLeft;
         }
         result.translate(tx, ty);
         return result;
